Rename dark mode button files
