+++ src/10_Styles/Main.scss
@@ -24,7 +24,7 @@
 @import "../03_Atoms/Avatar/Avatar";
 @import "../03_Atoms/Icon/Icon";
 @import "../03_Atoms/Typography/Typography";
-@import "../03_Atoms/DarkModeSwitch/DarkModeSwitch";
+@import "../03_Atoms/DarkModeButton/DarkModeButton";
 
 // Molecule Styles
 @import "../04_Molecules/Contact/Contact";
